requirements.txt
aiohttp
pyrofork
python-dotenv
tgcrypto
motor
aiofiles
dnspython
requests
jinja2
cloudscraper
psutil
speedtest-cli


nudenet
opencv-python-headless
